Split loader into query/data modules and add --batch-rows
Architecture: matches src/ pattern of one-word concern files —
query.rs owns TypeQL `given`-clause parsing, data.rs owns CSV reading
and per-cell value parsing, main.rs is orchestration only. Helpers
return Result<_, String>; only main calls `fatal`.

Behavior: --batch-rows (default 1000) chunks the CSV rows; each batch
runs in its own write transaction via query_with_inputs and commits
before the next. A progress line is printed after each commit.
